Avance
se unieron los avances que se han realizado

+++ database/migrations/gesap/2016_06_29_110902_create_anteproyecto_table.php
@@ -19,7 +19,7 @@ class CreateAnteproyectoTable extends Migration
             $table->String('NPRY_Keywords', 300);
             $table->String('NPRY_Descripcion', 500);
             $table->integer('NPRY_Duracion');
-            $table->integer('FK_NPRY_Pre_Director')->unsigned();
+            $table->bigInteger('FK_NPRY_Pre_Director')->unsigned();
             $table->foreign('FK_NPRY_Pre_Director')
                 ->references('PK_User_Codigo')
                 ->on('TBL_Usuario')
